# schema/openapi.yaml
...
openapi: 3.0.3
info:
  title: Swiss Stage API
  description: >
    囲碁・将棋大会運営プラットフォームのREST API。
    レスポンスは統一フォーマット(success/data/meta または success/error)。
    IDはULID文字列、日時はISO8601文字列、JSONはcamelCase。
    更新系はversionによる楽観ロック(競合は409)。
  version: 1.0.0
servers:
  - url: /
    description: 同一オリジン(SPAと同居)

# デフォルトは運営者のセッションCookie認証。公開エンドポイントは operation 側で security: [] を明示
security:
  - sessionCookie: []
paths:
...
  /api/v1/tournaments/{id}/teams:
    parameters:
      - $ref: "#/components/parameters/TournamentId"
    get:
      operationId: listTeams
      summary: チーム一覧(エントリー順)
      responses:
        "200":
          $ref: "#/components/responses/TeamList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createTeam
      summary: チーム作成(大会開始後は409)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          $ref: "#/components/responses/Team"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/tournaments/{id}/teams/csv-import:
    post:
      operationId: importTeams
      summary: チーム+メンバー一覧のCSVインポート(UTF-8/Shift_JIS自動判定)
      parameters:
        - $ref: "#/components/parameters/TournamentId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: >-
                    ヘッダー行必須(チーム名,氏名,段級位,ポジション[,グループ])。
                    連続する同一チーム名の行を1チームとして扱う。上限500行・1MB
      responses:
        "201":
          $ref: "#/components/responses/TeamCsvImportResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/tournaments/{id}/teams/csv-export:
    get:
      operationId: exportTeams
      summary: >-
        チーム+メンバー一覧CSVダウンロード(CSVインポートと同じ列構成。チーム名,氏名,段級位,ポジション,グループ。
        メンバー1人につき1行、エントリー順。0件時はヘッダー行のみでテンプレートとして使える。
        メンバーが1人もいないチームは行として出力されない。大会の状態を問わず利用可)
      parameters:
        - $ref: "#/components/parameters/TournamentId"
      responses:
        "200":
          $ref: "#/components/responses/TeamCsvExport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/tournaments/{id}/teams/{tid}:
    parameters:
      - $ref: "#/components/parameters/TournamentId"
      - $ref: "#/components/parameters/TeamId"
    patch:
      operationId: updateTeam
      summary: チーム更新(棄権・グループ割当変更を含む。nullの項目は変更しない)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
      responses:
        "200":
          $ref: "#/components/responses/Team"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteTeam
      summary: チーム削除(大会開始前のみ。開始後は409)
      responses:
        "204":
          description: 削除済み
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/tournaments/{id}/teams/{tid}/members:
    parameters:
      - $ref: "#/components/parameters/TournamentId"
      - $ref: "#/components/parameters/TeamId"
    post:
      operationId: addTeamMember
      summary: チームメンバー追加(役割・棋力込み。大会開始後は409)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTeamMemberRequest"
      responses:
        "201":
          $ref: "#/components/responses/Team"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/tournaments/{id}/teams/{tid}/members/{memberId}:
    parameters:
      - $ref: "#/components/parameters/TournamentId"
      - $ref: "#/components/parameters/TeamId"
      - $ref: "#/components/parameters/TeamMemberId"
    patch:
      operationId: updateTeamMember
      summary: チームメンバー更新(nullの項目は変更しない)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamMemberRequest"
      responses:
        "200":
          $ref: "#/components/responses/Team"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteTeamMember
      summary: チームメンバー削除(大会開始前のみ。開始後は409)
      responses:
        "200":
          $ref: "#/components/responses/Team"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: swiss_stage_session
      description: Google OAuth2ログイン後に発行されるJWTセッションCookie(運営者)
  parameters:
    TournamentId:
      name: id
      in: path
      required: true
      description: 大会ID(ULID)
      schema:
        $ref: "#/components/schemas/Ulid"
...
    TeamId:
      name: tid
      in: path
      required: true
      description: チームID(ULID)
      schema:
        $ref: "#/components/schemas/Ulid"
    TeamMemberId:
      name: memberId
      in: path
      required: true
      description: チームメンバーID(ULID)
      schema:
        $ref: "#/components/schemas/Ulid"
...
  responses:
...
    Team:
      description: チーム
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [success, data, meta]
            properties:
              success:
                type: boolean
                enum: [true]
              data:
                $ref: "#/components/schemas/Team"
              meta:
                $ref: "#/components/schemas/Meta"
    TeamList:
      description: チーム一覧
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [success, data, meta]
            properties:
              success:
                type: boolean
                enum: [true]
              data:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
              meta:
                $ref: "#/components/schemas/Meta"
    TeamCsvImportResult:
      description: チームCSVインポート結果
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [success, data, meta]
            properties:
              success:
                type: boolean
                enum: [true]
              data:
                $ref: "#/components/schemas/TeamCsvImportResult"
              meta:
                $ref: "#/components/schemas/Meta"
    TeamCsvExport:
      description: チーム+メンバー一覧CSV(UTF-8 BOM付き。チーム名,氏名,段級位,ポジション,グループ。メンバー1人につき1行)
      content:
        text/csv:
          schema:
            type: string
            format: binary
...
    BadRequest:
      description: バリデーションエラー・入力不正
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 未認証
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: 未存在(他人のリソースを含む。存在を漏らさない)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: 競合・状態遷移違反(楽観ロック・確定済み等)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    Ulid:
      type: string
      description: ULID(26文字のCrockford Base32)
      pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
    Meta:
      type: object
      additionalProperties: false
      required: [timestamp]
      properties:
        timestamp:
          type: string
          description: サーバー時刻(ISO8601)
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: "#/components/schemas/ApiErrorBody"
    ApiErrorBody:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          description: エラーコード(06_error_handling_design.md の一覧に従う)
        message:
          type: string
          description: そのままUIに表示できる日本語メッセージ
        details:
          type: array
          description: バリデーションエラー時のみ(空なら出力しない)
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field:
          type: string
          description: フィールド名(CSVインポートでは「n行目」)
        reason:
          type: string

    # ---- enum ----
...
    ParticipantStatus:
      type: string
      enum: [ACTIVE, WITHDRAWN]
...
    Rank:
      type: string
      nullable: true
      description: >
        棋力(段級位)。29段階。弱→強: KYU_20 < … < KYU_1 < DAN_1(初段)< … < DAN_9。
        順序はordinalではなくsortOrder(段=負・級=正・小さいほど強い)で判定する。
        null = 未入力(このenumを参照する全フィールドがnull許容のため、ここでnullableを定義)。
      enum:
        - KYU_20
        - KYU_19
        - KYU_18
        - KYU_17
        - KYU_16
        - KYU_15
        - KYU_14
        - KYU_13
        - KYU_12
        - KYU_11
        - KYU_10
        - KYU_9
        - KYU_8
        - KYU_7
        - KYU_6
        - KYU_5
        - KYU_4
        - KYU_3
        - KYU_2
        - KYU_1
        - DAN_1
        - DAN_2
        - DAN_3
        - DAN_4
        - DAN_5
        - DAN_6
        - DAN_7
        - DAN_8
        - DAN_9
...
    TeamMember:
      type: object
      additionalProperties: false
      description: >
        boardPosition: 1..teamSizeの整数(必須ポジション。1=主将,2=副将,3=三将,4=四将,5=五将)、
        またはnull(補欠)。対局結果はこのboardPositionではなくボード位置単位で記録され、
        実際に誰(正メンバーか補欠か)が対局したかは対局結果に紐付けない(オーダー管理はしない)
      required: [id, name, rank, boardPosition]
      properties:
        id:
          $ref: "#/components/schemas/Ulid"
        name:
          type: string
          maxLength: 50
        rank:
          $ref: "#/components/schemas/Rank"
        boardPosition:
          type: integer
          nullable: true
          minimum: 1
          maximum: 5
    Team:
      type: object
      additionalProperties: false
      required: [id, name, entryOrder, status, groupId, members]
      properties:
        id:
          $ref: "#/components/schemas/Ulid"
        name:
          type: string
          maxLength: 50
        entryOrder:
          type: integer
          minimum: 1
          description: チーム登録順。団体戦の初回ラウンドはこの順のみでペアリングする(棋力シードなし)
        status:
          $ref: "#/components/schemas/ParticipantStatus"
        groupId:
          $ref: "#/components/schemas/Ulid"
        members:
          type: array
          items:
            $ref: "#/components/schemas/TeamMember"
...
    CreateTeamRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        groupId:
          type: string
          nullable: true
          pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
          description: 省略時は先頭グループ(定義順)に割当
    UpdateTeamRequest:
      type: object
      additionalProperties: false
      description: nullの項目は変更しない。groupId は割当先グループの変更(PREPARING中のみ)
      properties:
        name:
          type: string
          maxLength: 50
        groupId:
          $ref: "#/components/schemas/Ulid"
        status:
          $ref: "#/components/schemas/ParticipantStatus"
    AddTeamMemberRequest:
      type: object
      additionalProperties: false
      description: >
        boardPosition省略(null)は補欠として追加する。1..teamSizeの範囲外・重複指定は400、
        補欠人数の上限超過は400
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        rank:
          $ref: "#/components/schemas/Rank"
        boardPosition:
          type: integer
          nullable: true
          minimum: 1
          maximum: 5
    UpdateTeamMemberRequest:
      type: object
      additionalProperties: false
      description: >
        nullの項目は変更しない。補欠に戻す場合は clearBoardPosition=true
        (boardPositionとの同時指定は400)
      properties:
        name:
          type: string
          maxLength: 50
        rank:
          $ref: "#/components/schemas/Rank"
        clearRank:
          type: boolean
        boardPosition:
          type: integer
          minimum: 1
          maximum: 5
        clearBoardPosition:
          type: boolean
    TeamCsvImportResult:
      type: object
      additionalProperties: false
      required: [importedTeamCount, teams]
      properties:
        importedTeamCount:
          type: integer
        teams:
          type: array
          description: 全行正常時のみ取り込む(エラー時は400 + details)
          items:
            $ref: "#/components/schemas/Team"
...